refactor: move looseEqual to vModel
